third_party/WebKit/LayoutTests/plugins/no-mime-with-valid-extension.html

<script src="../resources/plugin.js"></script>
<script>
if (window.testRunner)
    testRunner.dumpAsText();
startAfterLoadAndFinish();
</script>
<p>
This test checks that <a href="https://bugs.webkit.org/show_bug.cgi?id=50657">bug 50568</a>
is fixed. logargs is used to log initialization of the plugin.  If the bug is
present, the logargs attribute that triggers the log will be lost and nothing
will be displayed on screen. Upon success, this test should log a message at
initialization.
</p>
<!-- Embed tag with missing type="" parameter -->
<embed id="plugin" name="plugin" src="resources/test.blinktestplugin" logargs="1">
